chat-bot-cm-onboard-backend/pyproject.toml
[project]
name = "chat-bot-cm-onboard"
version = "0.1.0"
description = "Chatbot backend using FastAPI + LangChain + OpenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "sqlalchemy>=2.0.43",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.11.0",
    "langchain-openai>=0.3.35",
    "openai>=2.2.0",
    "langchain>=0.3.27",
    "alembic>=1.16.5",
]

[tool.setuptools]
packages = ["app", "alembic"]

#[tool.uv.scripts]
#server = "uvicorn app.main:app --reload"
